+++ apps/product/src/features/entry/components/inspector/fields/index.ts
@@ -8,4 +8,3 @@ export { TimeRow } from './TimeRow';
 // タグ・メモ・アラート
 export { NoteSection } from './NoteSection';
 export { TagRow } from './TagRow';
-export { TimeConflictAlert } from './TimeConflictAlert';
