+++ apps/product/src/features/entry/components/inspector/hooks/index.ts
@@ -1,4 +1,2 @@
-export { useInspectorNavigation } from './useInspectorNavigation';
-
 // re-export from feature hooks (used by EntryInspector.tsx)
 export { useInspectorKeyboard } from '../../../hooks/useInspectorKeyboard';
