+++ apps/product/src/features/entry/domain/index.ts
@@ -1,9 +1,7 @@
 export { determineEntryOrigin, isPlannedEntry, isUnplannedEntry } from './entry-time-model';
-export type { EntryLike } from './entry-time-model';
 export { findSkippableAutoRecords } from './skippable-auto-records';
 
 export { aggregateDayOfWeekDistribution } from './day-of-week-distribution';
-export { buildTimeUpdateData, buildUndoTimeUpdateData } from './entry-time-update';
 export {
   aggregatePlanLogEstimationAccuracy,
   transformEstimationAccuracy,
