+++ apps/product/src/features/entry/hooks/index.ts
@@ -1,5 +1,4 @@
 // Custom Hooks
-export { useEntryMutations } from './useEntryMutations';
 export { useFindSkippableAutoRecords } from './useFindSkippableAutoRecords';
 export { useTimeModelRecordMutations } from './useTimeModelRecordMutations';
 export { useTimeModelWriteMutations } from './useTimeModelWriteMutations';
