+++ apps/product/src/features/entry/index.ts
@@ -19,7 +19,6 @@ export type { PlanEvent, PlanEventStatus } from './types/plan-event';
 // Hooks
 // =============================================================================
 export {
-  useEntryMutations,
   useFindSkippableAutoRecords,
   useTimeModelRecordMutations,
   useTimeModelWriteMutations,
@@ -39,18 +38,8 @@ export { entryTintColor } from './lib/entry-tint';
 // =============================================================================
 // Domain (Entry 時間モデル — 純粋関数、DB/tRPC/React 非依存)
 // =============================================================================
-export {
-  buildTimeUpdateData,
-  buildUndoTimeUpdateData,
-  isPlannedEntry,
-  isUnplannedEntry,
-} from './domain';
-export type { EntryLike } from './domain';
-export {
-  isPlanRecordDrop,
-  isPlanTimeEditable,
-  resolveTimeModelDestination,
-} from './domain/time-model-destination';
+export { isPlannedEntry, isUnplannedEntry } from './domain';
+export { isPlanRecordDrop, resolveTimeModelDestination } from './domain/time-model-destination';
 export type { TimeModelDestination } from './domain/time-model-destination';
 
 // =============================================================================
@@ -81,7 +70,7 @@ export type { EntryCardPosition } from './components/card';
 // =============================================================================
 // Components (time model 記録導線)
 // =============================================================================
-export { ConfirmDayButton, RecordPlanButton } from './components/time-model/TimeModelRecordActions';
+export { ConfirmDayButton } from './components/time-model/TimeModelRecordActions';
 
 // =============================================================================
 // Components (Inspector fields — 他 feature から再利用可能な入力 row)
